requirements.txt
streamlit
llama-cpp-python
pandas
torch
transformers
sentence-transformers
chromadb
plotly
scikit-learn
